--- ArquivosCPP/Pilula.cpp
@@ -1,3 +1,0 @@
-#include "../ArquivosH/Pilula.h"
-
-// Implementação da classe pilula
