--- ArquivosCPP/Tijolo.cpp
@@ -1,3 +1,0 @@
-#include "../ArquivosH/Tijolo.h"
-
-// Implementação tijolos
