--- ArquivosH/Pilula.h
@@ -1,17 +1,0 @@
-#ifndef PILULA_H
-#define	PILULA_H
-
-// Classe de pilulas que herda da classe base EstruturasMapa
-
-class Pilula
-{
-private:
-	int altura;
-	int comprimento;
-public:
-	Pilula();
-	~Pilula();
-};
-
-#endif // !PILULAS_H
-
