--- ArquivosH/Tijolo.h
@@ -1,17 +1,0 @@
-#ifndef TIJOLO_H
-#define TIJOLO_H
-// Classe tijolo que herdará da classe base EstruturasMapa
-
-class Tijolo
-{
-private:
-	int Altura;
-	int Comprimento;
-public:
-	void Desenhar();
-	Tijolo();
-	~Tijolo();
-};
-
-
-#endif // !TIJOLO_H
